+++ components/catalogs/back/tests/Integration/Infrastructure/Controller/Public/GetProductUuidsActionTest.php
@@ -1,0 +1,143 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Controller\Public;
+
+use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
+use PHPUnit\Framework\Assert;
+use Ramsey\Uuid\Uuid;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+
+/**
+ * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @covers \Akeneo\Catalogs\Infrastructure\Controller\Public\GetProductUuidsAction
+ * @covers \Akeneo\Catalogs\Application\Handler\GetProductUuidsHandler
+ */
+class GetProductUuidsActionTest extends IntegrationTestCase
+{
+    private ?KernelBrowser $client = null;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->purgeDataAndLoadMinimalCatalog();
+    }
+
+    public function testItGetsPaginatedProductUuidsByCatalogId(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient(['read_catalogs', 'read_products']);
+        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi');
+        $this->enableCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c');
+        $this->createProduct('blue');
+        $this->createProduct('green');
+
+        $this->client->request(
+            'GET',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/product-uuids',
+            [
+                'limit' => 2,
+            ],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+        $payload = \json_decode($response->getContent(), true, 512, JSON_THROW_ON_ERROR);
+
+        Assert::assertEquals(200, $response->getStatusCode());
+        Assert::assertCount(2, $payload['_embedded']['items']);
+        foreach ($payload['_embedded']['items'] as $uuid) {
+            Assert::assertTrue(Uuid::isValid($uuid), 'Not a valid UUID');
+        }
+    }
+
+    public function testItReturnsAnEmptyListWhenTheCatalogIsDisabled(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient(['read_catalogs', 'read_products']);
+        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi');
+        $this->createProduct('blue');
+
+        $this->client->request(
+            'GET',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/product-uuids',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+        $payload = \json_decode($response->getContent(), true, 512, JSON_THROW_ON_ERROR);
+
+        Assert::assertEquals(200, $response->getStatusCode());
+        Assert::assertCount(0, $payload['_embedded']['items']);
+    }
+
+    public function testItReturnsBadRequestWhenPaginationIsInvalid(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient(['read_catalogs', 'read_products']);
+        $this->createCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c', 'Store US', 'shopifi');
+        $this->enableCatalog('db1079b6-f397-4a6a-bae4-8658e64ad47c');
+
+        $this->client->request(
+            'GET',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/product-uuids',
+            [
+                'limit' => -1,
+            ],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(422, $response->getStatusCode());
+    }
+
+    public function testItReturnsForbiddenWhenMissingPermissions(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient([]);
+
+        $this->client->request(
+            'GET',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/product-uuids',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(403, $response->getStatusCode());
+    }
+
+    public function testItReturnsNotFoundWhenCalalogDoesNotExist(): void
+    {
+        $this->client = $this->getAuthenticatedPublicApiClient(['read_catalogs', 'read_products']);
+
+        $this->client->request(
+            'GET',
+            '/api/rest/v1/catalogs/db1079b6-f397-4a6a-bae4-8658e64ad47c/product-uuids',
+            [],
+            [],
+            [
+                'CONTENT_TYPE' => 'application/json',
+            ],
+        );
+
+        $response = $this->client->getResponse();
+
+        Assert::assertEquals(404, $response->getStatusCode());
+    }
+}