+++ components/catalogs/back/tests/Integration/Infrastructure/Validation/CatalogUpdatePayloadTest.php
@@ -22,7 +22,11 @@ class CatalogUpdatePayloadTest extends IntegrationTestCase
     {
         parent::setUp();
 
+        $this->purgeDataAndLoadMinimalCatalog();
+
         $this->validator = self::getContainer()->get(ValidatorInterface::class);
+
+        $this->purgeDataAndLoadMinimalCatalog();
     }
 
     public function testItValidates(): void