+++ src/Akeneo/Category/back/tests/Specification/Infrastructure/Converter/InternalApi/InternalApiToStdSpec.php
@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 namespace Specification\Akeneo\Category\Infrastructure\Converter\InternalApi;
 
-use Akeneo\Category\Application\Converter\AttributeRequirementChecker;
-use Akeneo\Category\Application\Converter\FieldsRequirementChecker;
+use Akeneo\Category\Application\Converter\Checker\InternalApiRequirementChecker;
+use Akeneo\Category\Application\Converter\ConverterInterface;
+use Akeneo\Category\Infrastructure\Converter\InternalApi\InternalApiToStd;
 use PhpSpec\ObjectBehavior;
 
 /**
@@ -14,15 +15,18 @@ use PhpSpec\ObjectBehavior;
  */
 class InternalApiToStdSpec extends ObjectBehavior
 {
-    public function let(
-        FieldsRequirementChecker $fieldsRequirementChecker,
-        AttributeRequirementChecker $attributeChecker
-    ): void
+    public function let(InternalApiRequirementChecker $checker): void
     {
-        $this->beConstructedWith($fieldsRequirementChecker, $attributeChecker);
+        $this->beConstructedWith($checker);
     }
 
-    public function it_converts()
+    function it_is_initializable(): void
+    {
+        $this->shouldHaveType(InternalApiToStd::class);
+        $this->shouldImplement(ConverterInterface::class);
+    }
+
+    public function it_converts($checker)
     {
         $data = [
             'properties' => [
@@ -60,6 +64,7 @@ class InternalApiToStdSpec extends ObjectBehavior
                 ]
             ]
         ];
+        $checker->check($data)->shouldBeCalled();
         $this->convert($data)->shouldReturn($expected);
     }
 }