+++ src/Akeneo/Category/front/src/feature/components/EditAttributesForm.tsx
@@ -1,22 +1,23 @@
-import React, {useCallback, useState} from 'react';
-import {Field, FileInfo, MediaFileInput, SectionTitle, TextAreaInput, TextInput, Helper} from 'akeneo-design-system';
-import {Locale, LocaleSelector, useTranslate} from '@akeneo-pim-community/shared';
-import {useTemplate} from '../hooks';
+import React, {useCallback, useState, useMemo} from 'react';
 import styled from 'styled-components';
+
+import {SectionTitle, Helper} from 'akeneo-design-system';
+import {Locale, LocaleSelector, useTranslate} from '@akeneo-pim-community/shared';
+
+import {useTemplate} from '../hooks';
 import {
   Attribute,
   buildCompositeKey,
-  CategoryAttributes,
-  CategoryAttributeValueWrapper,
-  Template,
-} from '../models';
-
-import {
-  CategoryAttributeDefinition,
   CategoryAttributeValueData,
-  CategoryImageAttributeValueData,
   EnrichCategory,
+  isCategoryImageAttributeValueData,
 } from '../models';
+import {attributeFieldFactory} from './attributes/templateAttributesFactory';
+import {AttributeInputValue, buildDefaultAttributeInputValue, isImageAttributeInputValue} from './attributes/types';
+import {
+  convertCategoryImageAttributeValueDataToFileInfo,
+  convertFileInfoToCategoryImageAttributeValueData,
+} from '../helpers';
 
 const locales: Locale[] = [
   {
@@ -42,11 +43,6 @@ interface Props {
   ) => void;
 }
 
-const dumbUploader = async (file: File, onProgress: (ratio: number) => void) => ({
-  filePath: 'foo',
-  originalFilename: 'bar',
-});
-
 const FormContainer = styled.div`
   margin-top: 20px;
 
@@ -55,35 +51,32 @@ const FormContainer = styled.div`
   }
 `;
 
-const Field960 = styled(Field)`
-  max-width: 960px;
-`;
-
 export const EditAttributesForm = ({attributeValues, onAttributeValueChange}: Props) => {
   const [locale, setLocale] = useState('en_US');
   const translate = useTranslate();
 
   const handleTextChange = useCallback(
-    (attribute: Attribute) => (value: string) => {
+    (attribute: Attribute) => (value: AttributeInputValue) => {
+      if (isImageAttributeInputValue(value)) {
+        return;
+      }
       onAttributeValueChange(attribute, locale, value);
     },
     [locale, onAttributeValueChange]
   );
 
   const handleImageChange = useCallback(
-    (attribute: Attribute) => (value: FileInfo | null) => {
+    (attribute: Attribute) => (value: AttributeInputValue) => {
+      if (!isImageAttributeInputValue(value)) {
+        return;
+      }
+
       // TODO handle value===null
       if (!value || !value.size || !value.mimeType || !value.extension) {
         return;
       }
-      const data: CategoryImageAttributeValueData = {
-        size: value.size,
-        file_path: value.filePath,
-        mime_type: value.mimeType,
-        extension: value.extension,
-        original_filename: value.originalFilename,
-      };
-      onAttributeValueChange(attribute, locale, data);
+
+      onAttributeValueChange(attribute, locale, convertFileInfoToCategoryImageAttributeValueData(value));
     },
     [locale, onAttributeValueChange]
   );
@@ -97,20 +90,65 @@ export const EditAttributesForm = ({attributeValues, onAttributeValueChange}: Pr
     [attributeValues, locale]
   );
 
-  const {data: template, isLoading, isError, error} = useTemplate('02274dac-e99a-4e1d-8f9b-794d4c3ba330');
+  // TODO change hardcoded value to use the template uuid
+  const {data: template, isLoading, isError} = useTemplate('02274dac-e99a-4e1d-8f9b-794d4c3ba330');
+
+  const handlers = useMemo(() => {
+    const handlersMap: {[attributeUUID: string]: (value: AttributeInputValue) => void} = {};
+    template?.attributes.forEach((attribute: Attribute) => {
+      handlersMap[attribute.code] =
+        attribute.type === 'image' ? handleImageChange(attribute) : handleTextChange(attribute);
+    });
+    return handlersMap;
+  }, [template, handleImageChange, handleTextChange]);
 
   if (isLoading) {
-    return null; //TODO
+    return <h1>LOADING ...</h1>;
   }
 
   if (isError) {
-    console.log(error); //TODO
-    return (
-      <Helper level="error">
-        {error?.message}
-      </Helper>
-    );
+    return <Helper level="error">{translate('akeneo.category.edition_form.template.fetching_failed')}</Helper>;
   }
+
+  let attributesByOrder: Attribute[] = [];
+  template?.attributes.forEach((attribute: Attribute) => {
+    attributesByOrder[attribute.order] = attribute;
+  });
+
+  const attributeFields = attributesByOrder.map((attribute: Attribute) => {
+    const AttributeField = attributeFieldFactory(attribute);
+
+    if (AttributeField === null) {
+      return <Helper level="error">Could not find builder for {attribute.type} </Helper>;
+    }
+
+    const effectiveLocaleCode = attribute.is_localizable ? locale : null;
+    const compositeKey = buildCompositeKey(attribute, effectiveLocaleCode);
+
+    let value = attributeValues[compositeKey];
+
+    let dataForInput;
+    if (value) {
+      let {data: dataFromModel} = value;
+
+      if (isCategoryImageAttributeValueData(dataFromModel)) {
+        dataForInput = convertCategoryImageAttributeValueDataToFileInfo(dataFromModel);
+      } else {
+        dataForInput = dataFromModel;
+      }
+    } else {
+      dataForInput = buildDefaultAttributeInputValue(attribute.type);
+    }
+
+    return (
+      <AttributeField
+        locale={locale}
+        value={dataForInput}
+        onChange={handlers[attribute.code]}
+        key={attribute.uuid}
+      ></AttributeField>
+    );
+  });
 
   return (
     <FormContainer>
@@ -119,48 +157,7 @@ export const EditAttributesForm = ({attributeValues, onAttributeValueChange}: Pr
         <SectionTitle.Spacer />
         <LocaleSelector value={locale} values={locales} onChange={setLocale} />
       </SectionTitle>
-
-      {template?.attributes.map((attribute: Attribute) => {
-        const value = getAttributeValues(attribute);
-
-        switch (attribute.type) {
-          case 'text':
-            return (
-              <Field key={attribute.uuid} label={attribute.labels[locale]} locale={locale}>
-                <TextInput name={attribute.code} value={typeof value === 'string' ? value : ''} onChange={handleTextChange(attribute)} />
-              </Field>
-            );
-          case 'richtext':
-            return (
-              <Field960 key={attribute.uuid} label={attribute.labels[locale]} locale={locale}>
-                <TextAreaInput isRichText name={attribute.code} value={typeof value === 'string' ? value : ''} onChange={handleTextChange(attribute)} />
-              </Field960>
-            );
-          case 'textarea':
-            return (
-              <Field key={attribute.uuid} label={attribute.labels[locale]} locale={locale}>
-                <TextAreaInput name={attribute.code} value={typeof value === 'string' ? value : ''} onChange={handleTextChange(attribute)} />
-              </Field>
-            );
-          case 'image':
-            return (
-              <Field key={attribute.uuid} label={attribute.labels[locale]}>
-                <MediaFileInput
-                  value={null}
-                  onChange={handleImageChange(attribute)}
-                  placeholder="Drag and drop to upload or click here"
-                  uploadingLabel="Uploading..."
-                  uploadErrorLabel="An error occurred during upload"
-                  clearTitle="Clear"
-                  thumbnailUrl={null}
-                  uploader={dumbUploader}
-                />
-              </Field>
-            );
-        }
-
-        return null;
-      })}
+      {attributeFields}
     </FormContainer>
   );
 };