+++ src/Akeneo/Category/front/src/feature/models/Attribute.ts
@@ -3,6 +3,7 @@ import {LabelCollection} from '@akeneo-pim-community/shared';
 export type CategoryAttributeUUID = string;
 export type CategoryAttributeCode = string;
 
+// TODO keep or remove ?
 export const CATEGORY_ATTRIBUTE_TYPE_TEXT = 'text';
 export const CATEGORY_ATTRIBUTE_TYPE_IMAGE = 'image';