+++ src/Akeneo/Category/front/src/feature/models/CompositeKey.ts
@@ -1,5 +1,5 @@
 import {LocaleCode} from '@akeneo-pim-community/shared';
-import {Attribute, CategoryAttributeDefinition} from './Attribute';
+import {Attribute} from './Attribute';
 
 // a composite key points on some attribute value in a category data structure
 // it is of the form "<attribute code>|<attribute UUID>" with eventual suffix "|<locale>" if the attribute is localizable
@@ -10,10 +10,7 @@ export type CompositeKeyWithoutLocale = string;
 
 const COMPOSITE_KEY_SEPARATOR = '|';
 
-export function buildCompositeKey(
-  attribute: Attribute | CategoryAttributeDefinition,
-  localeCode: LocaleCode | null = null
-): CompositeKey {
+export function buildCompositeKey(attribute: Attribute, localeCode: LocaleCode | null = null): CompositeKey {
   const {code, uuid} = attribute;
   const components = [code, uuid];